+++ texecom_alarm/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.2] - 2026-09-01
+
 ### Added
 
 - On start, the add-on logs its version at the default log level so a Supervisor